sdefl buffer alloc fix

+++ Shaders/Bitpack.hlsl
@@ -98,7 +98,7 @@ f16_4 UnpackColor4UintF16(uint color)
         f16((color >> 0u)  & 0xFFu),
         f16((color >> 8u)  & 0xFFu),
         f16((color >> 16u) & 0xFFu),
-        f16((color >> 24u) & 0xFFu)) * (1.0f / 255.0f);
+        f16((color >> 24u) & 0xFFu)) * f16(1.0f / 255.0f);
 }
 
 f16_3 UnpackColor3Uint(uint color)
